Update to Python 3.13.9
Updated project to use Python 3.13.9. Chose this version
at this time because it has been released for a year and
end of support is 2029 (4 years from now).

General changes to support update:
- Updated various packages in pyproject.toml.
- Updated unit tests to use assertEqual (no "s") because
  assertEquals was removed in Python 3.11.
- Ran "uv sync" to update uv.lock file.

Stats/notebook changes to support update:
- df.groupby('col1')['col2'] now returns a multi-index.
  Fixed instances in analysis stats to drop multi-index
  levels before combining back to the original df.
- Fixed cannot convert nan to integer in draft stats.
- Removed .style call to not need jinja dependency in
  draft stats.
- Removed ESPN adjust points notebook (broken, currently
  not used).

Testing:
- Unit tests pass by running "uv run -m unittest -v -b" from
  inside the tests folder
- Ran every notebook in espn_nhlapi_analysis in VS Code
  Jupyter interactive window
  -> Note: draft_goon_stats_notebook.py currently not working
     on newer datasets for separate reason from this update

+++ espn_nhlapi_analysis/draft/draft_age_notebook.py
@@ -20,8 +20,8 @@ for i, owner in enumerate(owners_list):
 
     # Stats
     mean = round(owner_df['Player Age'].mean(), 1)
-    min = int(owner_df['Player Age'].min())
-    max = int(owner_df['Player Age'].max())
+    min = float(owner_df['Player Age'].min())
+    max = float(owner_df['Player Age'].max())
 
     fig.add_trace(go.Histogram(x=owner_df['Player Age'], histnorm='percent', name=f"{owner} (min={min} avg={mean} max={max})"), 1, i + 1)
 fig.update_layout(title="Distribution of Owner's Draft Picks (Overall)", xaxis_title="Age", yaxis_title="% of Picks", width=1400, height=500)
@@ -36,8 +36,8 @@ for season, season_df in draft_df.groupby('Season'):
 
         # Stats
         mean = round(owner_df['Player Age'].mean(), 1)
-        min = int(owner_df['Player Age'].min())
-        max = int(owner_df['Player Age'].max())
+        min = float(owner_df['Player Age'].min())
+        max = float(owner_df['Player Age'].max())
 
         fig.add_trace(go.Histogram(x=owner_df['Player Age'], histnorm='percent', name=f"{owner} (min={min} avg={mean} max={max})"), 1, i + 1)
     fig.update_layout(title=f"Distribution of Owner's Draft Picks ({season})", xaxis_title="Age", yaxis_title="% of Picks", width=1400, height=500)
